Alternative idea for Metric-producing conduit
Throws an exception on bad data (which will drop the connection, probably
that's what you want).

Doesn't require the last metric to have a trailing newline, and in fact I think
it considers a trailing newline at the end of a connection to be an error. That
could be tidied up if needs be.